Metadata page is now a single page and can be used everywhere
